# docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.CategoryCreateRequest:
    properties:
      description:
        example: Electronic items
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
    required:
    - name
    type: object
  dtos.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        example: Electronic items
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
      updated_at:
        type: string
    type: object
  dtos.CategoryUpdateRequest:
    properties:
      description:
        example: Electronic items
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
    type: object
  dtos.CheckoutItem:
    properties:
      product_id:
        example: 1
        type: integer
      quantity:
        example: 2
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dtos.CheckoutItemResponse:
    properties:
      price:
        example: 10000000
        type: integer
      product_id:
        example: 1
        type: integer
      product_name:
        example: Laptop
        type: string
      quantity:
        example: 2
        type: integer
      subtotal:
        example: 20000000
        type: integer
    type: object
  dtos.CheckoutRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dtos.CheckoutItem'
        minItems: 1
        type: array
    required:
    - items
    type: object
  dtos.CheckoutResponse:
    properties:
      created_at:
        type: string
      items:
        items:
          $ref: '#/definitions/dtos.CheckoutItemResponse'
        type: array
      total_amount:
        example: 20000000
        type: integer
      transaction_id:
        example: 1
        type: integer
    type: object
  dtos.CheckoutSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.CheckoutResponse'
      message:
        example: Checkout completed successfully
        type: string
    type: object
  dtos.CreateCategorySuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.CategoryResponse'
      message:
        example: Category created successfully
        type: string
    type: object
  dtos.CreateProductSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.ProductResponse'
      message:
        example: Product created successfully
        type: string
    type: object
  dtos.DeleteCategorySuccessResponse:
    properties:
      message:
        example: Category deleted successfully
        type: string
    type: object
  dtos.DeleteProductSuccessResponse:
    properties:
      message:
        example: Product deleted successfully
        type: string
    type: object
  dtos.GetCategoriesSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.CategoryResponse'
        type: array
      message:
        example: Categories retrieved successfully
        type: string
      total:
        example: 1
        type: integer
    type: object
  dtos.GetCategorySuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.CategoryResponse'
      message:
        example: Category retrieved successfully
        type: string
    type: object
  dtos.GetProductSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.ProductResponse'
      message:
        example: Product retrieved successfully
        type: string
    type: object
  dtos.GetProductsSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.ProductResponse'
        type: array
      message:
        example: Products retrieved successfully
        type: string
      total:
        example: 1
        type: integer
    type: object
  dtos.ProductCreateRequest:
    properties:
      category_id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 10000000
        type: integer
      stock:
        example: 5
        type: integer
    required:
    - name
    - price
    - stock
    type: object
  dtos.ProductResponse:
    properties:
      category_id:
        example: 1
        type: integer
      category_name:
        example: Electronic
        type: string
      created_at:
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 10000000
        type: integer
      stock:
        example: 5
        type: integer
      updated_at:
        type: string
    type: object
  dtos.ProductUpdateRequest:
    properties:
      category_id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 10000000
        type: integer
      stock:
        example: 5
        type: integer
    type: object
  dtos.UpdateCategorySuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.CategoryResponse'
      message:
        example: Category updated successfully
        type: string
    type: object
  dtos.UpdateProductResponse:
    properties:
      category_id:
        example: 1
        type: integer
      created_at:
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 10000000
        type: integer
      stock:
        example: 5
        type: integer
      updated_at:
        type: string
    type: object
  dtos.UpdateProductSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.UpdateProductResponse'
      message:
        example: Product updated successfully
        type: string
    type: object
  kasir-api_dtos.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.kasir-api.com/support
  description: A complete Point of Sale (POS) API built with Go, Gin, and GORM
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kasir API
  version: "1.0"
paths:
  /categories:
    get:
      description: Retrieve a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetCategoriesSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category with the provided details
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateCategorySuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Delete a category by ID (soft delete)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.DeleteCategorySuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Delete a category
      tags:
      - Categories
    get:
      description: Retrieve a specific category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetCategorySuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Get a category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UpdateCategorySuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Update a category
      tags:
      - Categories
  /products:
    get:
      description: Retrieve a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetProductsSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided details
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Delete a product by ID (soft delete)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.DeleteProductSuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Delete a product
      tags:
      - Products
    get:
      description: Retrieve a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetProductSuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UpdateProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Update a product
      tags:
      - Products
  /products/search:
    get:
      description: Search products by name or category name
      parameters:
      - description: Search query for product name or category name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetProductsSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Search products
      tags:
      - Products
  /transactions/checkout:
    post:
      consumes:
      - application/json
      description: Process a checkout transaction with items and calculate total amount
      parameters:
      - description: Items to checkout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CheckoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CheckoutSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kasir-api_dtos.ErrorResponse'
      summary: Complete transaction checkout
      tags:
      - Transactions
schemes:
- http
swagger: "2.0"